backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==3.0.2
dj-database-url==2.1.0
Django==3.2.8
django-cors-headers==4.0.0
django-phonenumber-field==7.1.0
django-shortuuidfield==0.1.3
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
mysqlclient==2.1.1
phonenumbers==8.13.13
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
service-identity==21.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
zope.interface==6.0
